requirements.txt
paapi5-python-sdk==1.1.0
requests==2.32.3
PyYAML==6.0.2
Jinja2==3.1.4
pandas==2.2.2
python-slugify==8.0.4
